kill soundcloud downloader
